Add vmnet package
The package manages the vmnet-helper[1] child process, providing
connection to the vment network without running the guest as root.

We will use vment-helper for the vfkit driver, which does not have a way
to use shared network, when guests can access other guest in the
network.  We can use it later with the qemu driver as alternative to
socket_vment.

[1] https://github.com/nirs/vmnet-helper