krunkit: Add krunkit driver
krunkit is a tool to launch configurable virtual machines using the
libkrun platform, optimized for GPU accelerated virtual machines and AI
workloads on Apple silicon.

It is mostly compatible with vfkit; the driver is a simplified copy of
the vfkit driver. Unlike vfkit, krunkit is available only on Apple
silicon.

Changes compared to vfkit driver:
- krunkit requires unix socket for netwroking, so we must use
  vment-helper.
- krunkit can be controlled only via HTTP, not via unix socket.
- krunkit does not support --kernel-cmdline
- We must enable vmnet offloading, required for krunkit.
- The code was simplified since vmnet-helper is always used
- Code was cleaned up to use .ResolveStorePath()

Limitations:
- Only one machine can be created since we use the same port for krunkit
  --restful-uri. This should be fixed to use an unused port, or use a
  unix socket when unix socket is supported[1].

[1] https://github.com/containers/krunkit/issues/47