iso: Fix console for vfkit/krunkit
The serial console name depends on the driver. We had setting for qemu
that does not work for vfkit and krunkit, likely hiding serial console
output during boot.

We use 2 console= configurations, one is known to work. I hope that with
this we can have one iso image compatible with all drivers.