iso: Add default configs
Graceful shutdown using --restul-uri does not work with vfkit and
krunkit with the new 6.6 kernel. We know that it works with kernel built
with default architecture ocnfig. Adding all default configs fixes this
issue.

Change was generate by:

1. creating a default config for arm64:

    cd out/buildroot/output-aarch64/build/linux-6.6.95
    make ARCH=arm64 defconfig
    make ARCH=arm64 menuconfig
    (disable platfrom support and exit)

2. Copy generated defconfig to our linux defconfig

    cd out/buildroot/output-aarch64
    make linux-update-defconfig
    (edit: remove deletion, keep additions)

3. Run menuconfig to normalzie the linux defconfig

   make linux-menuconfig-aarch64

The additonl configs increase the kernel size from 388M to 393M.

I'm sure we can remove most of the added configs, but I don't knwo which
configs solved this issue. We can try to minimize the image later.