drivers: Add VirtioFS shares for vfkit and krunkit
Add --virtiofs-share flag for specifying one or more shared directories
on the host to share with the guest using Virtiofs.

Testing with krunkit shows that virtiofs mount is 10 times faster than
9p mount.

Usage:

   --virtiofs-share HOST_PATH:GUEST_PATH

The value can include multiple shares separated by "," or repeated
multiple times, once for every share.

Example:

    minikube start --virtiofs-share ~/models:/mnt/models

The arguments are parsed and validated when configuring the driver, so
invalid arguments fail quickly without starting the driver. The
validated shares are stored in the machine config:

    $ jq '.Driver.VirtiofsShares' < ~/.minikube/machines/minikube/config.json
    [
      {
        "HostPath": "/Users/joe/models",
        "GuestPath": "/mnt/models",
        "MountTag": "f845b54d-00e3-493d-9541-3b37490b96db"
      }
    ]

Minikube generates a new random UUID for every virtiofs share to
identify the file system inside the guest. In krunkit and vfkit, every
share will be added as:

    --device virtio-fs,sharedDir=/host-path,mountTag=f845b54d-00e3-493d-9541-3b37490b96db

When the guest is started the share is mounted via SSH using:

    sudo mkdir -p /mnt/models
    sudo mount -t virtiofs f845b54d-00e3-493d-9541-3b37490b96db /mnt/models

Example mount:

    $ minikube ssh findmnt /mnt/models
    TARGET      SOURCE                               FSTYPE   OPTIONS
    /mnt/models f845b54d-00e3-493d-9541-3b37490b96db virtiofs rw,relatime

More work is needed to add VirtioFS to qemu. I'm starting with fkit and
krunkit since they have identical interface (krunkit was designed as
drop-in replacement for vfkit).