kvm: Don't mess with console.log
We used to create ~/.minikube/machines/NAME/console.log as the user
running the machine, so ensure that minikube can dump the log contents
when we wait for IP address. This breaks minikube on Fedora and RHEL,
since slinux does not allow libvirt to access the log file when it was
not created properly by libvirt.

Having better logs in the kvm driver can be helpful but it cannot break
basic functionality. Remove the code to create the log file and let
libvirt create the log file in the expected way.

We also removed the log file after starting the cluster which is very
wrong. Libvirt continue to write the removed file, and possibly helpful
logs can never be accessed.