kvm: Convert to internal driver
Remove the docker-machine-driver-kvm2 wrapper and use the kvm driver as
internal driver.

To avoid dependency on libvirt shared library on Linux, we build now
with libvirt_dlopen tag. With this we don't need libvirt devel libraies
during build, and in runtime we will fail if libvirt shared library is
not installed.

May need more work on checking if libvirt shared library is installed
and recommending installing libvirt to use the driver.