Try to debug macos smoke tests
Disable the skip on high load since the load does not seem related to
the failures. We get the same failures with high or minimal load.

Run top for 60 seconds before the test to let the runner cool down after
the build tasks. Top output will help to understand the load on the
runner.

Run top for 10 seconds after the tests to understand the load when the
test completed or failed.

Remove the stop and second start, we can add this when starting minikube
works reliably.

Tweak minikube arguments to match vment-helper integration tests, known
to work reliably:
- cpus: 1
- memory: 2600m