drivers: Run drivers in the forground
We were running qemu with -daemonize and vfkit, and krunkit in a new
process group. This is similar to using the `&` operator in the shell to
run a process in the background.

Running in the background is useful when the user tries to terminate the
entire minikube process group using `killpg`, and want to keep the vm
process running. This is feature is used by ramen testing tool to
terminate child processes after an error.

In macOS 15, Apple introduced the *Local network privacy*[1], limiting
access to the local network. When an untrusted application (like
minikube, vfkit, krunkit, qemu, and vmnet-helper) is trying to access
the local network, the user is asked to approve the operation.
Approvals can be also managed via *System Settings -> Privacy & Security
-> Local Network*. There is no way to manage approvals programmatically,
only via the UI.

On GitHub actions, we run in a headless server, so there is no way to
allow access to local network. Apple mitigated this by automatically
allowing local network access to command-line tools run over SSH,
including any child processes they spawn.

However running a child process as a daemon (qemu) or in a new process
group (vfkit, krunkit) seems to break the relation between the sshd
process and child process, so network operations are blocked in the
child process. When this happens, the network operation will fail with
*no route to host*, as we see in GitHub CI:

    I1217 19:10:10.289490   28945 main.go:143] libmachine: Waiting until SSH server "192.168.105.2:22" is accessible
    I1217 19:10:10.289497   28945 main.go:143] libmachine: Dialing to "192.168.105.2:22"
    I1217 19:10:10.290521   28945 main.go:143] libmachine: Failed to dial: dial tcp 192.168.105.2:22: connect: no route to host
    I1217 19:10:11.291543   28945 main.go:143] libmachine: Dialing to "192.168.105.2:22"
    I1217 19:10:11.291773   28945 main.go:143] libmachine: Failed to dial: dial tcp 192.168.105.2:22: connect: no route to host
    I1217 19:10:12.292106   28945 main.go:143] libmachine: Dialing to "192.168.105.2:22"
    I1217 19:10:12.292373   28945 main.go:143] libmachine: Failed to dial: dial tcp 192.168.105.2:22: connect: no route to host
    ...
    I1217 19:14:56.609203   28945 main.go:143] libmachine: Failed to dial: dial tcp 192.168.105.2:22: connect: no route to host
    I1217 19:14:56.705836   28945 start.go:128] duration metric: took 6m0.010647057s to createHost
    W1217 19:14:56.705915   28945 start.go:715] error starting host: creating host: create host timed out in 360.000000 seconds

To fix the issue, we run qemu, vfkit, krunkit, and vment-helper in the
foreground, and they inherit the local network access privileges from
the sshd process.

This may cause issues to users assuming the previous behavior so we may
need to make this configurable. I'm starting with simple change to test
if this fixes the issues in GitHub actions.

[1] https://developer.apple.com/documentation/technotes/tn3179-understanding-local-network-privacy