Inspect networking state before and after
Since we solved the load issue, minimize top output to 2 samples with 10
seconds delay before and after the test.

Log output of ifconfig before and after the test. This will show if the
bridge is created by vmnet when starting the vm.

Log routing table before and after the test. This will show if the
bridge got the expected IP (192.168.105.1) and if there is a conflict
with another interface using the same IP.